modularize and bug fixes
